@@ -22,7 +22,7 @@ def setup_package():
         description="Python Checksumming Library",
         # in pkg-info this maps to 'description'
         long_description="Intuitive and extendable checksumming for python objects",
-        python_requires='>=3.5',
+        python_requires='>=3.8',
         keywords="checksum checksumming hashing",
         tests_require=['pytest'],
         zip_safe=False,
@@ -38,10 +38,11 @@ def setup_package():
             "Operating System :: OS Independent",
             "Programming Language :: Python",
             "Programming Language :: Python :: 3",
-            "Programming Language :: Python :: 3.5",
-            "Programming Language :: Python :: 3.6",
-            "Programming Language :: Python :: 3.7",
             "Programming Language :: Python :: 3.8",
+            "Programming Language :: Python :: 3.9",
+            "Programming Language :: Python :: 3.10",
+            "Programming Language :: Python :: 3.11",
+            "Programming Language :: Python :: 3.12",
             "Topic :: Scientific/Engineering",
         ]
     )
